% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load_dot_env}
\alias{load_dot_env}
\title{Load environment variables from the specified file}
\usage{
load_dot_env(file = ".env")
}
\arguments{
\item{file}{The name of the file to load from.}
}
\description{
Load environment variables from the specified file
}

